BK-1862 Make Security module more extendable
