+++ .gitignore
@@ -13,3 +13,5 @@
 setup.cfg
 .env
 .idea
+
+tests/static/css
